refactor(LottoPrize): 로또 상금 타입을 BigDecimal로 변경
나중에 확장성을 위해 로또 상금 타입을 BigDecimal로 변경했습니다.

+++ src/main/java/lotto/data/LottoResult.java
@@ -29,7 +29,7 @@ public class LottoResult {
         BigDecimal totalPrize = BigDecimal.ZERO;
         for (LottoPrize lottoResult : lottoRank.keySet()) {
             BigDecimal count = lottoRank.get(lottoResult);
-            BigDecimal prize = BigDecimal.valueOf(lottoResult.getPrize());
+            BigDecimal prize = lottoResult.getPrize();
             totalPrize = totalPrize.add(prize.multiply(count));
         }
         return totalPrize;
